Move track functions to specific file
