+++ src/OpenLoco/openloco.vcxproj
@@ -72,6 +72,7 @@
     <ClCompile Include="Map\Tile.cpp" />
     <ClCompile Include="Map\TileManager.cpp" />
     <ClCompile Include="Map\Track\SubpositionData.cpp" />
+    <ClCompile Include="Map\Track\Track.cpp" />
     <ClCompile Include="Map\Track\TrackData.cpp" />
     <ClCompile Include="Map\WaveManager.cpp" />
     <ClCompile Include="Math\Trigonometry.cpp" />
@@ -281,6 +282,7 @@
     <ClInclude Include="Map\TileLoop.hpp" />
     <ClInclude Include="Map\TileManager.h" />
     <ClInclude Include="Map\Track\SubpositionData.h" />
+    <ClInclude Include="Map\Track\Track.h" />
     <ClInclude Include="Map\Track\TrackData.h" />
     <ClInclude Include="Map\Wave.h" />
     <ClInclude Include="Map\WaveManager.h" />
